backend/requirements.txt
# Padosi Politics - Python Dependencies
# Flask Backend Requirements

# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Authentication & Security
Flask-Security-Too==5.3.3
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
passlib==1.7.4

# API & Validation
Flask-CORS==4.0.0
marshmallow==3.20.1

# Performance & Caching
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Database Migrations
Flask-Migrate==4.0.5
alembic==1.13.1

# Email (Optional)
Flask-Mail==0.9.1

# Background Tasks (Optional - Celery)
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2

# Production Server
gunicorn==21.2.0

# Date/Time handling
python-dateutil==2.8.2

# PostgreSQL driver (for production)
psycopg2-binary==2.9.9
